pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["UMESH519 <umesh.karamchandani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
debugpy = "1.6.4"
ipykernel = "^6.20.2"
jupyter = "^1.0.0"
pandas = "^1.5.3"
pyarrow = "^10.0.1"
tqdm = "^4.64.1"
plotly = "^5.13.0"
scikit-learn = "^1.2.1"
xgboost = "^1.7.3"
cmake = "^3.25.0"
lightgbm = "3.3.3"
pip = "^23.0"
python-dotenv = "^0.21.1"
optuna = "^3.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
